+++ straxen/plugins/events/__init__.py
@@ -60,3 +60,9 @@ from .local_minimum_info import *
 
 from . import multi_scatter
 from .multi_scatter import *
+
+from . import event_bipo_variables
+from .event_bipo_variables import *
+
+from . import event_bipo_matching
+from .event_bipo_matching import *
